The Files
all of the project files. Clicking being functions relating the the placing of buy and sell orders along with the trading algorithm. clicks being the screen coordinates needed for each in game item. iodmiokf being the file that contacts the api and condenses the data. absolutes are something important but I forget what.